Massive refactor, move functions to appropriate modules/packages
